requirements.txt
# Core
numpy>=1.21.0
scipy>=1.7.0
librosa>=0.9.0
soundfile>=0.10.0
pyyaml>=6.0

# Deployment (Cloud/Render)
fastapi>=0.110.0
uvicorn>=0.27.0
jinja2>=3.1.0
websockets>=12.0
python-multipart>=0.0.9
gunicorn>=21.2.0
# tflite-runtime is better for cloud (small footprint)
# Users can install tensorflow locally if they want
tflite-runtime>=2.14.0

# Optional/Development
# tensorflow>=2.10.0
# matplotlib>=3.5.0
# seaborn>=0.12.0
# tqdm>=4.64.0
# kaggle>=1.5.0
# python-dotenv>=0.21.0
# pytest>=7.0.0
